+++ AlchemistCode/Config/AlchemistModConfig.cs
@@ -36,6 +36,10 @@ public class AlchemistModConfig : SimpleModConfig
 
     [ConfigSection("Accessibility")]
     [ConfigHoverTip]
+    public static bool ShowPoisonForecast { get; set; } = true;
+
+    [ConfigSection("Accessibility")]
+    [ConfigHoverTip]
     [ConfigColorPicker(EditAlpha = false)]
     public static Color AntitoxinBarColor { get; set; } = new("9B5CFF");